+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Allowed to pack RPM and DEB packages with ``cartridge pack``
+  comamnd without default dependencies file.
+
 ## [2.9.0] - 2021-04-26
 
 ### Changed