+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.9.1] - 2021-05-13
+
 ### Fixed
 
 - Allowed to pack RPM and DEB packages with ``cartridge pack``